BugFix. Si la db esta caida no borra el fichero de log.

+++ bin/ModSecProcessor.py
@@ -143,9 +143,9 @@ def ProcessFiles(hilo, fichero):
     
     query = "INSERT INTO log (file_id, host, get, date, cookie, referer, ua, xff, ae, message, timestamp) VALUES (\'" + mdb.escape_string(file_name) + "\', \'" + mdb.escape_string(host) + "\', \'" + mdb.escape_string(get) + "\', \'" + mdb.escape_string(date) + "\', \'" + mdb.escape_string(cookie) + "\', \'" + mdb.escape_string(referer) + "\', \'" + mdb.escape_string(ua) + "\', \'" + mdb.escape_string(xff) + "\', \'" + mdb.escape_string(ae) + "\', \'" + mdb.escape_string(message) + "\', \'" + str(ts) + "\');"
 
-    InsertDB(query)
-    #print "\t\t\t[-] Borrando: " + fichero
-    os.remove(fichero)
+    if (InsertDB(query)):
+        #print "\t\t\t[-] Borrando: " + fichero
+        os.remove(fichero)
     
 def main():
